+++ src/common/metadata-utils/index.ts
@@ -1,10 +1,12 @@
+import { SfdxError } from '@salesforce/core';
+import * as c from 'chalk';
 import * as child from 'child_process';
 import * as extractZip from 'extract-zip';
 import * as fs from 'fs-extra';
 import * as path from 'path';
 import * as sfdx from 'sfdx-node';
 import * as util from 'util';
-import { execSfdxJson, filterPackageXml } from '../../common/utils';
+import { execCommand, execSfdxJson, filterPackageXml, uxLog } from '../../common/utils';
 import { CONSTANTS } from '../../config';
 const exec = util.promisify(child.exec);
 
@@ -149,8 +151,34 @@ class MetadataUtils {
 
   // List installed packages on a org
   public static async listInstalledPackages(orgAlias: string = null, commandThis: any): Promise<any[]> {
-    const alreadyInstalled = await execSfdxJson('sfdx force:package:installed:list', commandThis, { fail: true });
+    let listCommand = 'sfdx force:package:installed:list';
+    if (orgAlias != null) {
+      listCommand += ` -u ${orgAlias}`;
+    }
+    const alreadyInstalled = await execSfdxJson(listCommand, commandThis, { fail: true });
     return alreadyInstalled?.result || [];
+  }
+
+  // Install package on existing org
+  public static async installPackagesOnOrg(packages: any[], orgAlias: string = null, commandThis: any = null) {
+    const alreadyInstalled = await MetadataUtils.listInstalledPackages(null, this);
+    for (const package1 of packages) {
+      if (alreadyInstalled.filter((installedPackage: any) =>
+        package1.SubscriberPackageVersionId === installedPackage.SubscriberPackageVersionId).length === 0) {
+        uxLog(commandThis, `Installing package ${package1.SubscriberPackageName} ${package1.SubscriberPackageVersionName}`);
+        if (package1.SubscriberPackageVersionId == null) {
+          throw new SfdxError(c.red(`[sfdx-hardis] You must define ${c.bold('SubscriberPackageVersionId')} in .sfdx-hardis.yml (in installedPackages property)`));
+        }
+        const securityType = package1.SecurityType || 'AllUsers' ;
+        let packageInstallCommand = `sfdx force:package:install --package ${package1.SubscriberPackageVersionId} --noprompt --securitytype ${securityType} -w 60`;
+        if (orgAlias != null) {
+          packageInstallCommand += ` -u ${orgAlias}`;
+        }
+        await execCommand(packageInstallCommand, this, { fail: true, output: true });
+      } else {
+        uxLog(commandThis, `Skip installation of ${package1.SubscriberPackageName} as it is already installed`);
+      }
+    }
   }
 
   // Retrieve metadatas from a package.xml
@@ -158,15 +186,15 @@ class MetadataUtils {
                                         filteredMetadatas: string[], options: any = {}, commandThis: any, debug: boolean) {
 
     // Build package.xml for all org
-    commandThis.ux.log(`[sfdx-hardis] Generating full package.xml from ${commandThis.org.getUsername()}...`);
+    uxLog(commandThis, `[sfdx-hardis] Generating full package.xml from ${commandThis.org.getUsername()}...`);
     const manifestRes = await exec('sfdx sfpowerkit:org:manifest:build -o package.xml');
     if (debug) {
-      commandThis.ux.log(manifestRes.stdout + manifestRes.stderr);
+      uxLog(commandThis, manifestRes.stdout + manifestRes.stderr);
     }
 
     // Filter managed items if requested
     if (options.filterManagedItems) {
-      commandThis.ux.log('[sfdx-hardis] Filtering managed items from package.Xml manifest');
+      uxLog(commandThis, '[sfdx-hardis] Filtering managed items from package.Xml manifest');
       // List installed packages & collect managed namespaces
       const installedPackages = await this.listInstalledPackages(null, commandThis);
       const namespaces = [];
@@ -186,16 +214,16 @@ class MetadataUtils {
         removeFromPackageXmlFile: packageXmlToRemove,
         updateApiVersion: CONSTANTS.API_VERSION
       });
-      commandThis.ux.log(filterNamespaceRes.message);
+      uxLog(commandThis, filterNamespaceRes.message);
     }
 
     // Filter package XML to remove identified metadatas
     const filterRes = await filterPackageXml(packageXml, packageXml, { removeMetadatas: filteredMetadatas });
-    commandThis.ux.log(filterRes.message);
+    uxLog(commandThis, filterRes.message);
 
     // Retrieve metadatas
     if (fs.readdirSync(metadataFolder).length === 0 || checkEmpty === false) {
-      commandThis.ux.log(`[sfdx-hardis] Retrieving metadatas in ${metadataFolder}...`);
+      uxLog(commandThis, `[sfdx-hardis] Retrieving metadatas in ${metadataFolder}...`);
       const retrieveRes = await sfdx.mdapi.retrieve({
         retrievetargetdir: metadataFolder,
         unpackaged: packageXml,
@@ -205,10 +233,10 @@ class MetadataUtils {
         _rejectOnError: true
       });
       if (debug) {
-        commandThis.ux.log(retrieveRes);
+        uxLog(commandThis, retrieveRes);
       }
       // Unzip metadatas
-      commandThis.ux.log('[sfdx-hardis] Unzipping metadatas...');
+      uxLog(commandThis, '[sfdx-hardis] Unzipping metadatas...');
       await extractZip(path.join(metadataFolder, 'unpackaged.zip'), { dir: metadataFolder });
       await fs.unlink(path.join(metadataFolder, 'unpackaged.zip'));
     }