+++ src/common/utils/index.ts
@@ -17,6 +17,8 @@ if (isGitRepo()) {
 
 let pluginsStdout = null;
 
+export const isCI = process.env.CI != null;
+
 export function isGitRepo() {
   const isInsideWorkTree = child.spawnSync(
     'git',
@@ -97,7 +99,7 @@ export async function execCommand(
     debug: false
   }
 ): Promise<any> {
-  commandThis.ux.log(`[sfdx-hardis][command] ${c.bold(c.grey(command))}`);
+  uxLog(commandThis, `[sfdx-hardis][command] ${c.bold(c.grey(command))}`);
   let commandResult = null;
   // Call command (disable color before for json parsing)
   const prevForceColor = process.env.FORCE_COLOR;
@@ -119,7 +121,7 @@ export async function execCommand(
   }
   // Display output if requested, for better user unrstanding of the logs
   if (options.output || options.debug) {
-    commandThis.ux.log(`[sfdx-hardis][commandresult] ${commandResult.stdout}`);
+    uxLog(commandThis, `[commandresult] ${commandResult.stdout}`);
   }
   // Return status 0 if not --json
   process.env.FORCE_COLOR = prevForceColor;
@@ -279,8 +281,8 @@ export async function catchMatches(
         catcherLabel
       });
       if (commandThis.debug) {
-        commandThis.ux.log(
-          `[sfdx-hardis] [${fileName}]: Match [${matches}] occurences of [${catcher.type}/${catcher.name}] with catcher [${catcherLabel}]`
+        uxLog(commandThis,
+          `[${fileName}]: Match [${matches}] occurences of [${catcher.type}/${catcher.name}] with catcher [${catcherLabel}]`
         );
       }
     }
@@ -331,11 +333,53 @@ export async function generateReports(
     columns
   });
   await fs.writeFile(reportFileExcel, excel, 'utf8');
-  commandThis.ux.log('[sfdx-hardis] Generated report files:');
-  commandThis.ux.log(`[sfdx-hardis] - CSV: ${reportFile}`);
-  commandThis.ux.log(`[sfdx-hardis] - XLS: ${reportFileExcel}`);
+  uxLog(commandThis, 'Generated report files:');
+  uxLog(commandThis, `- CSV: ${reportFile}`);
+  uxLog(commandThis, `- XLS: ${reportFileExcel}`);
   return [
     { type: 'csv', file: reportFile },
     { type: 'xls', file: reportFileExcel }
   ];
+}
+
+export function uxLog(commandThis: any, text: string) {
+  text = (text.includes('[sfdx-hardis]')) ? text : '[sfdx-hardis]' + (text.startsWith('[') ? '' : ' ') + text;
+  if (commandThis?.ux) {
+    commandThis.ux.log(text);
+  } else {
+    console.log(text);
+  }
+}
+
+// Caching methods
+const SFDX_LOCAL_FOLDER = '/root/.sfdx';
+const TMP_COPY_FOLDER = '.cache/sfdx-hardis/.sfdx';
+let RESTORED = false;
+
+// Put local sfdx folder in tmp/sfdx-hardis-local for CI tools needing cache/artifacts to be within repo dir
+export async function copyLocalSfdxInfo() {
+  if (!isCI) {
+    return;
+  }
+  if (fs.existsSync(SFDX_LOCAL_FOLDER)) {
+    await fs.ensureDir(path.dirname(TMP_COPY_FOLDER));
+    await fs.copy(SFDX_LOCAL_FOLDER, TMP_COPY_FOLDER, { dereference: true , overwrite: true });
+    // uxLog(this, `[cache] Copied sfdx cache in ${TMP_COPY_FOLDER} for later reuse`);
+    // const files = fs.readdirSync(TMP_COPY_FOLDER, {withFileTypes: true}).map(item => item.name);
+    // uxLog(this, '[cache]' + JSON.stringify(files));
+  }
+}
+
+// Restore only once local Sfdx folder
+export async function restoreLocalSfdxInfo() {
+  if ((!isCI) || RESTORED === true) {
+    return;
+  }
+  if (fs.existsSync(TMP_COPY_FOLDER)) {
+    await fs.copy(TMP_COPY_FOLDER, SFDX_LOCAL_FOLDER, { dereference: true, overwrite: false });
+    // uxLog(this, '[cache] Restored cache for CI');
+    // const files = fs.readdirSync(SFDX_LOCAL_FOLDER, {withFileTypes: true}).map(item => item.name);
+    // uxLog(this, '[cache]' + JSON.stringify(files));
+    RESTORED = true;
+  }
 }