+++ src/hooks/init/check-local-sfdx-hardis-files.ts
@@ -1,34 +1,41 @@
 import * as c from 'chalk';
 import * as fs from 'fs-extra';
+import * as prompts from 'prompts';
+import { isCI } from '../../common/utils';
 
 export const hook = async (options: any) => {
     // Skip hooks from other commands than hardis:scratch commands
     const commandId = options?.id || '';
-    // No await there because it can be processed while other commands
-    // tslint:disable no-floating-promises
-    managePackageJson(commandId);
-    manageGitIgnore(commandId);
-    // tslint:enable no-floating-promises
 
+    await managePackageJson(commandId);
+    await manageGitIgnore(commandId);
 };
 
 // Add utility scripts if they are not present
 async function managePackageJson(commandId: string) {
-    if (!commandId.startsWith('hardis:scratch')) {
+    if (!commandId.startsWith('hardis:scratch') && !commandId.startsWith('hardis:project:configure')) {
         return;
     }
     const packageJsonFile = './package.json';
     if (fs.existsSync(packageJsonFile)) {
         // Update existing package.json to define sfdx utility scripts
-        fs.readFile(packageJsonFile, 'utf8').then(async (text: string) => {
-            const packageJson = JSON.parse(text);
-            const hardisPackageJsonContent = await getSfdxHardisPackageJsonContent();
-            packageJson['scripts'] = Object.assign(hardisPackageJsonContent['scripts'], packageJson['scripts']);
-            if (JSON.stringify(packageJson) !== JSON.stringify(JSON.parse(text))) {
+        const text = await fs.readFile(packageJsonFile, 'utf8');
+        const packageJson = JSON.parse(text);
+        const hardisPackageJsonContent = await getSfdxHardisPackageJsonContent();
+        packageJson['scripts'] = Object.assign(packageJson['scripts'], hardisPackageJsonContent['scripts']);
+        if (JSON.stringify(packageJson) !== JSON.stringify(JSON.parse(text)) && !isCI) {
+            const confirm = await prompts({
+                type: 'confirm',
+                name: 'value',
+                initial: true,
+                message: 'Your package.json is deprecated, do you agree to upgrade it ? (If you hesitate, just trust us and accept)'
+            });
+            if (confirm.value === true) {
                 await fs.writeFile(packageJsonFile, JSON.stringify(packageJson, null, 2));
                 console.log(c.cyan('[sfdx-hardis] Updated package.json with sfdx-hardis content'));
             }
-        });
+        }
+
     } else {
         // Create package.json to define sfdx utility scripts
         const hardisPackageJsonContent = await getSfdxHardisPackageJsonContent();
@@ -53,23 +60,32 @@ async function manageGitIgnore(commandId: string) {
                 updated = true;
             }
         }
-        if (updated) {
-            await fs.writeFile(gitIgnoreFile, gitIgnoreLines.join('\n') + '\n', 'utf-8');
-            console.log(c.cyan('[sfdx-hardis] Updated .gitignore'));
+        if (updated && !isCI) {
+            const confirm = await prompts({
+                type: 'confirm',
+                name: 'value',
+                initial: true,
+                message: 'Your .gitignore is deprecated, do you agree to upgrade it ? (If you hesitate, just trust us and accept)'
+            });
+            if (confirm.value === true) {
+                await fs.writeFile(gitIgnoreFile, gitIgnoreLines.join('\n') + '\n', 'utf-8');
+                console.log(c.cyan('[sfdx-hardis] Updated .gitignore'));
+            }
         }
-
     }
 }
 
 async function getSfdxHardisPackageJsonContent() {
     const hardisPackageJsonContent = {
         scripts: {
-            'scratch:create': 'sfdx hardis:scratch:create',
             'scratch:push': 'sfdx force:source:push -g -w 60 --forceoverwrite',
             'scratch:pull': 'sfdx force:source:pull --forceoverwrite',
             'scratch:open': 'sfdx force:org:open',
             'org:test:apex': 'sfdx hardis:org:test:apex',
-            'login:reset': 'sfdx auth:logout --noprompt || true && sfdx config:unset defaultusername defaultdevhubusername -g && sfdx config:unset defaultusername defaultdevhubusername || true'
+            'scratch:create': 'sfdx hardis:scratch:create',
+            'login:reset': 'sfdx auth:logout --noprompt || true && sfdx config:unset defaultusername defaultdevhubusername -g && sfdx config:unset defaultusername defaultdevhubusername || true',
+            'configure:auth:devhub': 'sfdx hardis:project:configure:auth --devhub',
+            'configure:auth:deployment': 'sfdx hardis:project:configure:auth'
         }
     };
     return hardisPackageJsonContent;
@@ -77,6 +93,7 @@ async function getSfdxHardisPackageJsonContent() {
 
 async function getHardisGitRepoIgnoreContent() {
     const gitIgnoreContent = [
+        '.cache/',
         'config/user/',
         'hardis-report/',
         'tmp/'