+++ src/hooks/postrun/notify.ts
@@ -9,6 +9,7 @@ export const hook = async (options: any) => {
     if (!commandId.startsWith('hardis')) {
         return;
     }
+
     // Send hook to microsoft ?teams if MS_TEAMS_WEBHOOK_URL env var is set, or msTeamsWebhookUrl in config
     const config = await getConfig('user');
     const msTeamsWebhookUrl = process.env.MS_TEAMS_WEBHOOK_URL || config.msTeamsWebhookUrl ;